Pivot to having a Conn class
I didn't want to do this, but having explored lots of other things it feels like the best way to go.

Having a single sql function was a cool idea but would have been trouble when mixed with pooling / vfs / transactions.

Just too many things need access to the sqlite3 ptr.  And at least as written we can update the pointer contained inside the Conn which means commands are still possible.