Incremental Blob IO
Re-addingin incremental blob io via a custom sql function blob_io(ReadableStream/WritableStream, rowid, table_name, column_name, ?db_name) -> number of bytes read from or written to the SQLite blob.

When passing a ReadableStream, the bytes will be *read* from the stream and *written* to the SQLite Blob.

When passing a WritableStream, the bytes will be *written* to the stream, and *read* from the SQLite Blob.

This means that ReadableStreams cause database mutation, while passing a WritableStream does a database read.  This is a little counter intuitive, but I believe that it is the most ergonomic way of implementing this functionality.

[I think this blob io code is ok, but I think I've discovered some memory corruption on the VFS names maybe?]