Remove unnecessary async
The way my asyncify wrapper works is that returning a promise from an import causes a stack unwind / rewind.  If necessary, it may also allocate a stack to perform the unwinding.

Because of this, it's important to only return a promise if the underlying VFS returns a promise.  In particular, functions xFileSize and xFullPathname, were always async before even when most VFSs won't have async implementations of those functions.