Roughing out local VFS
The plan is to replace the OPFS and Picker VFSs with a single Local VFS that can mount any kind of FileSystemHandle (FileSystemFileHandle and FileSystemDirectoryHandle) into a single "tree".
The OPFS is just a FileSystemDirectoryHandle that you can access without prompting for user permission.
Another benefit of this is that journal files will naturally fit inside this mount / "tree" interface.  If you mount a FileSystemFileHandle into a sub-path of a FileSystemDirectoryHandle then the journal file will naturally be created inside that directory.  This means you could choose to put journal files for showFilePicker files into the Origin Private File System for instance.